NetworkAgent: Log a metric if a message is queued before registration
It should not be valid to queue messages before
registering a NetworkAgent, but the current design
makes it possible in practice. Log a metric so
that we have a handle on whether or not anybody
is doing this in the field.

Bug: 296355968
Test: manual
(cherry picked from https://android-review.googlesource.com/q/commit:0b1a7b8c98d30241925b4ce4475fa01ad4f763ed)
Merged-In: I83ea92e39a860cebe58d1d82cb024313e244b9aa
Change-Id: I83ea92e39a860cebe58d1d82cb024313e244b9aa